requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
pytorch-lightning>=2.0.0
lightly>=1.4.0
numpy>=1.21.0
scipy>=1.7.0
h5py>=3.7.0
Pillow>=9.0.0
tifffile>=2023.7.10
optuna>=3.0.0
plotly>=5.0.0
kaleido>=0.2.1